feat:created aux and reb ECU files
