+++ aux/src/app.c
@@ -1,5 +1,6 @@
 #include <app.h>
 #include <ecu.h>
+#include <ecu_aux.h>
 #include <mcal.h>
 
 uint8_t application_init()
@@ -10,7 +11,7 @@ uint8_t application_init()
         show_error("mcal_init FAIL\n");
     }
 
-    if(can_init() == FAIL)
+    if (can_init() == FAIL)
     {
         show_error("can_init FAIL\n");
     }
@@ -61,69 +62,6 @@ uint8_t hazard_lights_blink()
     }
 }
 
-uint8_t monitor_engine_block()
-{
-    while (1)
-    {
-        if (block_engine() == FAIL)
-        {
-            show_error("app.block_engine FAIL\n");
-        }
-        go_sleep(1);
-        if (unblock_engine() == FAIL)
-        {
-            show_error("app.unblock_engine FAIL\n");
-        }
-        go_sleep(1);
-    }
-}
-
-uint8_t send_can_hazard_light()
-{
-    while (1)
-    {
-        // TESTE; precisa abstrair
-        uint8_t status;
-        if (read_pin_status(&status, 9) == FAIL)
-        {
-            return FAIL;
-        }
-
-        if (status == 1)
-        {
-            if (set_pin_status(0, 9) == FAIL)
-            {
-                return FAIL;
-            }
-            if (can_send_hazard_light(1) == FAIL)
-            {
-                show_error("app.can_send_hazard FAIL\n");
-                return FAIL;
-            }
-        }
-
-        if (read_pin_status(&status, 8) == FAIL)
-        {
-            return FAIL;
-        }
-
-        if (status == 1)
-        {
-            if (set_pin_status(0, 8) == FAIL)
-            {
-                return FAIL;
-            }
-            if (can_send_hazard_light(0) == FAIL)
-            {
-                show_error("app.can_send_hazard FAIL\n");
-                return FAIL;
-            }
-        }
-
-        go_sleep(1);
-    }
-}
-
 uint8_t monitor_read_can()
 {
     while (1)
@@ -132,30 +70,14 @@ uint8_t monitor_read_can()
         struct can_frame frame = {
             .can_id = 29, .can_dlc = 8, .data = {0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF}};
 
-        printf("cand ID read %X\n", frame.can_id);
         if (can_read_vcan0(&frame) != FAIL)
         {
-            if (frame.can_id == 0x7E0)
-            {
-                handle_tcu_can(frame.data);
-            }
+            // TODO
         }
         else
         {
             show_error("Error monitor_read_can\n");
             go_sleep(2);
         }
-    }
-}
-
-uint8_t initiation_start_reb()
-{
-    while (1)
-    {
-        if (start_reb() == FAIL)
-        {
-            show_error("app.start_reb FAIL\n");
-        }
-        go_sleep(1);
     }
 }
