+++ aux/src/ecu_aux.c
@@ -1,0 +1,23 @@
+#include <ecu.h>
+#include <mcal.h>
+#include <unistd.h>
+
+uint8_t get_hazard_button_status(uint8_t *status)
+{
+    if (read_pin_status(status, 1) == FAIL)
+    {
+        show_error("ECU.read_pin_status FAIL\n");
+        return FAIL;
+    }
+    return SUCCESS;
+}
+
+uint8_t set_hazard_light(uint8_t status)
+{
+    if (set_pin_status(status, 0) == FAIL)
+    {
+        show_error("ECU.set_pin_status FAIL\n");
+        return FAIL;
+    }
+    return SUCCESS;
+}
