+++ bsw/inc/ecu.h
@@ -10,14 +10,6 @@
 #define S_ON 0x01U
 #define S_OFF 0x00U
 
-uint8_t get_hazard_button_status(uint8_t *status);
-uint8_t set_hazard_light(uint8_t status);
 uint8_t read_console();
-uint8_t start_reb();
-uint8_t block_engine();
-uint8_t unblock_engine();
-uint8_t engine_block_status(uint8_t *status);
-uint8_t can_send_hazard_light(uint8_t status);
-uint8_t handle_tcu_can(unsigned char data[]);
 
 #endif
