+++ reb/src/app.c
@@ -1,5 +1,6 @@
 #include <app.h>
 #include <ecu.h>
+#include <ecu_reb.h>
 #include <mcal.h>
 
 uint8_t application_init()
@@ -29,38 +30,6 @@ uint8_t read_input()
     }
 }
 
-uint8_t hazard_lights_blink()
-{
-    while (1)
-    {
-        uint8_t status = 0;
-        if (get_hazard_button_status(&status) == FAIL)
-        {
-            show_error("get_hazard_button_status FAIL\n");
-            return FAIL;
-        }
-        if (status == S_ON)
-        {
-            if (set_hazard_light(S_ON) == FAIL)
-            {
-                show_error("set_hazard_light FAIL\n");
-                return FAIL;
-            };
-            go_sleep(1);
-            if (set_hazard_light(S_OFF) == FAIL)
-            {
-                show_error("set_hazard_light FAIL\n");
-                return FAIL;
-            };
-            go_sleep(1);
-        }
-        else
-        {
-            go_sleep(2);
-        }
-    }
-}
-
 uint8_t monitor_engine_block()
 {
     while (1)
@@ -82,7 +51,7 @@ uint8_t send_can_hazard_light()
 {
     while (1)
     {
-        // TESTE; precisa abstrair
+        //@TODO isso será removido OU alterado conforme a necessidade
         uint8_t status;
         if (read_pin_status(&status, 9) == FAIL)
         {
@@ -132,7 +101,6 @@ uint8_t monitor_read_can()
         struct can_frame frame = {
             .can_id = 29, .can_dlc = 8, .data = {0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF}};
 
-        printf("cand ID read %X\n", frame.can_id);
         if (can_read_vcan0(&frame) != FAIL)
         {
             if (frame.can_id == 0x7E0)
